dr: introduce the dr command
The "odf dr" command exposes selected ramenctl[1] commands that are
suitable to use with ODF.

The command support the following sub-commands:

- init: Create configuration file for dr commands
- test: Test disaster recovery flow with a tiny application

The command is using the ramenctl/cmd.RootCmd with modified name and
description, and selected sub commands.

We annotate the command to let ramenctl know that it is running as "odf
dr". The name is used in the configuration file.

Running "odf dr" commands is identical to running "ramenctl" command.
The configuration file is used to adapt the command to different
Kubernetes distribution (distro), storage providers (pvcSpecs), OCM
configuration (clsuterSet) and DR configuration (drPolicy).

Issues:

- ramenctl/pkg/build should contain the version and git commit hash for
  dr commands reports. This information seems to be missing in odf cli.
- The globals flags inherited from the RootCmd are ignored by the dr
  command. There seems to be no good way to fix this in ramenctl.

[1] https://github.com/RamenDR/ramenctl